interactive game teaching teens about phishing, strong passwords, and privacy/requirements.txt
streamlit==1.31.0
zxcvbn==4.4.28
openai==1.10.0
Pillow==10.0.0
reportlab==4.0.4